Task 1: Temperature Conversion from Celcius to Fahrenheit with just 4 element using NumPy and Mapping Function Task 2: Comparing the time for execution for smaller and larger datasets (100,000 and 1,000,000) to measure the efficiency and feasibility of Mapping Function and NumPy Arrays!
NumPy wins hands down with larger datasets!

A simple Mapping Function or List Comprehension is just as good for smaller datasets! 

When applying Python and Python libraries in real life applications, first understand the project and its datasets and then come to a conclusion regarding what features and concepts makes more sense!

For a small dataset project depending on what defines a small and a large dataset, the optimal solution is to apply just what the project demands and not necessarily the most advanced functions and features of Python or any other Programming Language. 

Here for 4 just 4 elements or under 100,000 datasets, NumPy being a far more advanced feature in Python makes zero sense in terms of business, scalability and user experience wise compared to Mapping Function or a List Comprehension!

The idea is to get more out of less or let's say applying the right solution not necessarily the most advanced solution or feature!